+++ app/src/examples/RuntimeTests/tests/animations/withTiming/easingSnapshots.test.tsx
@@ -18,8 +18,8 @@ import {
   render,
   useTestRef,
   wait,
-} from '../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
-import { Snapshots } from './snapshots/withTiming.snapshot';
+} from '../../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+import { Snapshots } from './withTiming.snapshot';
 
 const AnimatedComponent = ({
   easing,
@@ -143,7 +143,7 @@ describe('withTiming snapshots 📸, test EASING', () => {
   (
     [
       {
-        easing: Easing.back(4.5),
+        easing: Easing.back(4.75),
         message: 'Easing.back(4.75)',
         snapshotName: 'back1',
       },
@@ -161,9 +161,6 @@ describe('withTiming snapshots 📸, test EASING', () => {
       await wait(1200);
       const updates = updatesContainer.getUpdates();
       expect(updates).toMatchSnapshot(Snapshots[snapshotName]);
-      expect(updates).toMatchNativeSnapshots(
-        await updatesContainer.getNativeSnapshots()
-      );
     });
   });
 });
